Bug 1314807 - Change ownership of files in workspace; r=dustin
The other changesets for this bug worked on Try but failed on
autoland due to permissions failures in /home/worker/workspace.
Ensure the files have the proper permissions.

The path is a cache with the repo and build type keyed into it.
So there should be no problem with this change causing permissions
issues for other tasks.

--HG--
extra : amend_source : 1fa1c2f7a22afabd4b2e252a268b2f492238966c