Bug 1302192: ensure that only one option matches by-*; r=wcosta
If an item matches exactly, it is always selected.  This will catch the case
where multiple regular expressions match (the previous version would silently
pick one based on dictionary order).

I was worried this was occurring, but because exact matches items are selected,
it is not; still, best to be sure.

MozReview-Commit-ID: EtZBTr61SvE

--HG--
extra : rebase_source : 64300b8722709711d9be66f63f808f4e340532d9