Bug 1347569: don't add optmized tasks to task.dependencies; r=Callek
`task.dependencies` is part of the internal taskgraph calculations, so it must
only refer to tasks that are being created, and not to those which already
exist.  `task.task['dependencies']`, on the other hand, is what is given to
`queue.createTask` and expresses all dependencies.

MozReview-Commit-ID: GJ6JVj6JMBz

--HG--
extra : rebase_source : 553aec85134fe0e41c53f917327b13d1b66592aa