Bug 1355180: introduce run-on-projects: built-projects, and use it; r=jmaher
This test-only "run-on-projects" option indicates that the test should run on
exactly the projects where its parent build runs.  This avoids the situation
where a build is configured to *not* run on a project, but some wayward test
didn't get the memo and "pulls in" the build anyway.  This has been
particularly problematic with win64-asan/opt.

MozReview-Commit-ID: 8UP6FsXBohy

--HG--
extra : rebase_source : 237ea2898767da8e4561d1f6884e5add92a7bde2