Bug 1391427 - Execute python toolchain scripts with mach. r=glandium
Run scripts with a `.py` filename extension through `./mach python`
so the normal enviroment with in-tree python libraries is available.

This is helpful for the Rust upstream release download and repackaging
steps, which are more easily expressed in python than in an sh-based
build script like we use for clang and other tools.

Invocation of `mach python` on Windows-hosted generic workers fails
because of some missing environment pieces. For the purposes of this
bug we can just run the script for Windows targets in a docker-worker
so Python support was left unimplemented for generic workers.

MozReview-Commit-ID: 4XUQ1XrVBc2

--HG--
extra : rebase_source : df81d2da7e70fffb29e96377f16ab22def9e94e0