Bug 1391427 - Add script 'arguments' key to toolchain tasks. r=glandium
Add an optional 'arguments' key to the yaml description for toolchain
tasks. This is a list of strings to be passed to the script invocation.

This lets us set behaviour, e.g. selecting the version to build or
selecting targets from the task description instead of having to
hard-code those things in the build script itself. Where the same
script otherwise works for multiple configurations, that is easier
to update and simplifies supporting variants.

MozReview-Commit-ID: 30oJYnQaZ7A

--HG--
extra : rebase_source : bdd7bdc5f874d1329ff52d900cd1ac93df23c6dc