bug 1412690 - beetmover-cdns. r=rail
This patch adds the beetmover-cdns kind, and adds it to
`publish_fennec`.

This was the first non-buildbot-bridge, non-dummy relpro task, so this
needed a new transform.

This patch also updates the `previous_graph_kinds` and updates the
beetmover scopes in scriptworker.py.

MozReview-Commit-ID: 3rpkjuLjjXz

--HG--
extra : rebase_source : d491f1ce5d10ee8f2602698236106007e203c358