Bug 1414921 - Add --geckoProfile to capture profiles from talos tasks to |mach try fuzzy|, r=dustin
Enables |./mach try fuzzy --talos-profile|. This template only applies to talos
tasks. It also provides --geckoProfile for consistency with |mach try syntax|,
but I don't like this name so it's hidden from the help.

The 'talos-profile.yml' template is also very specific (only applies to Talos
tasks).  Ideally I'd like a general 'command.yml' template that just appends
arguments to the command for any arbitrary tasks. But then we'd need to invent
an expression syntax in try_task_config.json so we could make sure it only
applies to Talos. Then I thought rather than implement it for a specific
template, we should have a general way of doing this which could apply to any
and all of the templates.

Needless to say, it's a rabbit hole and something that's best left to a
follow-up so we don't delay this bug.

MozReview-Commit-ID: GhllZ7sr0ar

--HG--
extra : rebase_source : 1de4deecc2f73130904d7c95d4ff12f85883cd91