Bug 1446241: Set run-on-projects for upload-generated-sources and upload-symbols r=aki
If tasks still exist before the optimization phase, then any tasks these depend
on will also be scheduled. In particular for devedition on m-r, that means that
although the builds were being excluded from the target tasks due to the
build's run-on-projects settings, the upload-symbols and
upload-generated-sources tasks did exist in the target tasks, and so the builds
got re-added to the target tasks during graph optimization.

MozReview-Commit-ID: 1AWJuafULEE

--HG--
extra : rebase_source : 153653733e01433ae894a0ba4fd99228e0936024