Bug 1415868 - Remove support for register_action_task; r=jonasfj,tomprince
In a post-actions-as-hooks world, users will not have scopes to create tasks,
so this mode of action definition will not be possible. This is not currently
used from Treeherder (it links to
https://tools.taskcluster.net/tasks/<taskid>/interactive instead)

This drops support for the JSON-e-only interactive action; that action is not
currently used from treeherder, so that should have no impact for users.

MozReview-Commit-ID: 9i3POpjahAc

--HG--
extra : rebase_source : e6de03389a0c5c67d5332d2b1c97e1d4bf6a22d3