Bug 1473048 - Prevent code coverage builds from running when using try option syntax. r=ahal
This patch completely disables *ccov, and *jsdcov builds and tests when scheduling them through try option syntax as these build variations use a lot of resources and are rarely needed to be scheduled. The only way to schedule code coverage from now on will be with the |mach try fuzzy| tool.

Differential Revision: https://phabricator.services.mozilla.com/D3921

--HG--
extra : moz-landing-system : lando