Bug 1397847: Add supported API to get tasks of a given kind from a taskgraph; r=dustin
`mach artifact toolchain` gets task definitions from taskgraph, to get the
index path to find the artifacts at. Now that these index paths depend on the
digests of fetch tasks, those kinds need to be loaded as well. This adds a
supported API to get task definitions for a given kind, which loads all the
kind dependencies.

Differential Revision: https://phabricator.services.mozilla.com/D12161

--HG--
extra : moz-landing-system : lando