Bug 1471905: [taskgraph] Consistently include the cache digests of parent tasks in downstream cached tasks; r=dustin
There are several kinds that cache tasks based on the inputs that go into the task. Historically,
these inputs included the name of upstream tasks. This change these tasks to include the digest
of the upstream tasks.

This also bumps the version of the docker and toolchain as every digest is changed for them.

Differential Revision: https://phabricator.services.mozilla.com/D11949

--HG--
extra : moz-landing-system : lando