Bug 1568006: Add the project's taskgraph directory to the python path; r=mhentges,dustin
For mozilla-central, all the code related to taskgraph lives in
taskcluster/taskgraph. For the out-of-tree version, we want to be able to have
repository-specific code. The natural place for it to live is an a package
beside taskcluster/ci. Add that to python path, and provide some hooks for
adding to the various registries in taskgraph.

Differential Revision: https://phabricator.services.mozilla.com/D38910