Bug 1619795: Allow specifying index-search via run-using; r=jlorenzo
In Bug 1525218, a toolchain task was added to fenix, to pull minidump_stackwalk
from mozilla-central. This used the `index_search` transform. However, the
`toolchain` kind is treated specially by taskgraph (in handling `fetches`).

If we ever want to define fenix-specific toolchains, the `index-search`
transform will conflict with the toolchain transforms. By switching to using a
job type, both types of tasks can co-exist in the kind (toolchains also use job
type).

The immediate motivation for implementing this now is to be able to specify the
`toolchain-artifact` attribute on the tasks, to support using toolchains from
mozilla-central that don't have a `public/build/` prefix. The `index_search`
transform could be changed to pass through `attributes`, but this has the same
effect, and is future-proof.

Differential Revision: https://phabricator.services.mozilla.com/D65250