Bug 1571562 - Allow toolchains to use fetches from other toolchains. r=tomprince
This also allows toolchain tasks to use aliases via fetches, which they
currently aren't allowed via use_toolchain. There are more toolchains
now than there were when the restriction was added, and it might be
useful to be able to use aliases. The flip side is that there are some
risks involved with aliases, which is why the restriction was there in
the first place. Let's see how things play out.

Backports https://phabricator.services.mozilla.com/D40713 to taskgraph

Differential Revision: https://phabricator.services.mozilla.com/D69341