Bug 1721187 - Ensure we use a modern version of pip3 in decision images, r=aki
The previous version of pip didn't look for wheels by default which resulted in
source distributions needing to be compiled. Also it was version 9.0.1.. so we
should upgrade it regardless.

This pins pip using the compatible release operator (so generating new images
will pick up new patch versions but not minor ones).

Differential Revision: https://phabricator.services.mozilla.com/D120259