Bug 1733454 - [gecko sync] Make sure we re-use the same default parameters in --diff mode, r=aki
Previously we were defaulting the parameters option to '[None]'. This caused
the generator to re-compute the default values across revisions in --diff mode,
resulting in different parameters being used and a different resulting
taskgraph.

By instantiating the parameters up front in this case, we can ensure that the
same parameter set is being used across both revisions.

This is a sync of the changes that have already landed in mozilla-central.

Differential Revision: https://phabricator.services.mozilla.com/D127268