Bug 1746414 - Support specifying defaults with 'extend_parameters_schema', r=taskgraph-reviewers,jmaher
When a downstream repo uses 'extend_parameters_schema', there's no way for them
to specify which default values to use in the event we use strict=False.

This adds a new optional argument which is a function that returns a dictionary
of default values to add. It also moves the "core" defaults (previously in the
'Parameters._fill_defaults' function) to one of these default factories as
well.

We need to use functions because computing the default values can be expensive
(i.e if they require querying version control), and we don't want to incur that
penalty unless they are actually being used (strict=False).

Differential Revision: https://phabricator.services.mozilla.com/D134524