No Bug - Search parent directories when instantiating a Repository object, r=releng-reviewers,jmaher
The 'get_repository' function was previously only searching the current path
that gets passed in for a repository. This means that the root of the repository
*must* be passed in.

In parameters.py, the '_repo' helper function was passing in os.getcwd(), which
meant that if this function got invoked when the user's CWD != repo root, we'd
raise an exception. In practice, this was the root cause of builds being broken
in Gecko.

Since it's not easy for 'taskgraph' itself to tell what the repo root is, we
can't provide any better default than CWD here. Instead, let's modify the
'get_repository' function to search ancestor directories for a repo.

The 'ancestors' utility function was pulled from 'mozbuild.base' in Gecko.

Differential Revision: https://phabricator.services.mozilla.com/D135229