No Bug - Use GraphConfig.vcs_root to populate default parameters, r=releng-reviewers,jcristau
This fixes a bug in Gecko where we were running a command that uses taskgraph
outside of the source repo (and therefore we were failing to instantiate the
Repository object).

Note if the root dir is not "taskcluster/ci" we won't be able to detect the vcs
root. In this case, fallback to 'os.getcwd()' (the previous value).

Differential Revision: https://phabricator.services.mozilla.com/D135550