feat(cli): add ability to exclude keys from a graph (no bug), r=taskgraph-reviewers,aki
This can be used to simplify what is returned from the invocation. E.g,

    $ taskgraph full --exclude-key task.payload.command --exclude-key attributes

The above command will strip out the command in the task's payload, as well
as the top-level attributes.

This can be especially useful when used in conjunction with the --diff flag and
there are known differences between the two graphs that we'd like to suppress
(to make sure there aren't *other* differences in addition)

Differential Revision: https://phabricator.services.mozilla.com/D140099