docs: split the setup tutorial into two smaller tutorials (no bug), r=releng-reviewers,jcristau
Tutorials should ideally be small and accomplishable. The fact that the first
step of the tutorial is "Talk to your Taskcluster administrator" means that no
one looking to get started with Taskgraph for the first time will ever get past
this.

This refactors the tutorial so the local Taskgraph scaffolding stuff is first
and standalone. This way new users can get a feel for Taskgraph and generate
graphs locally, even if they can't yet submit anything to Taskcluster. Having
this initial tutorial be achievable is important if we want users to give
Taskgraph a chance.

The next tutorial will still require contacting Taskcluster administrators, but
at least they will be able to complete the first one.

Depends on D140284

Differential Revision: https://phabricator.services.mozilla.com/D140285