fix: hook up 'description' and 'if-dependencies' to Task object (no bug), r=taskgraph-reviewers,aki
I got so pre-occupied with syncing task.py between the two taskgraphs that I
forgot to actually add the transform / generator plumbing to get
'if-dependencies' and the description added to the Task object.

Luckily nothing should be broken by this omission. I'm going to call this a
'fix' as we advertised 'if-dependencies/description' in the CHANGELOG already,
so this is a fix to actually get them working.

Differential Revision: https://phabricator.services.mozilla.com/D140881