feat(task): fetch run-task from artifacts (RELENG-787) No bug r=taskgraph-reviewers,ahal
As we move projects to pull taskgraph from pypi, we are also removing references to taskgraph's source repository (and eventually move from HG to a git repo).
With this change, we provide run-task and fetch-content files as artifact of the decision task, and download them as needed, removing the need to pull them from a repository.

Differential Revision: https://phabricator.services.mozilla.com/D141255