ci: use a 'constraints' file to install packages for tests
Previously the tasks in CI would install everything that exists in
requirements/test.txt, which included a lot of unnecessary packages.

By instead passing in the file as a constraint via the `-c` option, we
only install the packages that are actually needed, while still
constraining them to the versions specified in the requirements.txt.