docs: refactor contributing docs out of README
A README should primarily be about telling readers what the repo is
about, as well as provide some links to additional docs / context.
Information about how to contribute doesn't need to be front and centre
as relatively few readers of the README will become actual contributors.
Instead this information should be in a dedicated CONTRIBUTING file.

This also adds a 'How It Works' section to the README and the
CODE_OF_CONDUCT file.