test: upload xml code coverage artifacts from unit task
Since we run multiple versions of Python, we need to be careful not to
clobber the coverage data between runs. This is accomplished via the
`--cov-append` flag, which updates the existing data file (left by a
previous run) rather than clobber it.

This also introduces a `clean` and `report` tox environment. The former
ensures we don't accidentally append to a data file left by a previous
run, and the latter generates the coverage report. Locally, this will
just run `coverage report` but a CI task can e.g change the format to
xml and save it to a file for upload to codecov.io.