refactor!: rename taskgraph.util.WHITELISTED_SCHEMA_IDENTIFIERS
BREAKING CHANGE: taskgraph.util.WHITELISTED_SCHEMA_IDENTIFIERS is now
taskgraph.util.EXCEPTED_SCHEMA_IDENTIFIERS. Instead of a list of

In Gecko there's a lint check to prevent the use of
'whitelist/blacklist', and there's an exception for this case.

Since we've now landed gearing up to release Taskgraph 2.0 anyway, might as well
make the switch.

In addition to a callable, exceptions can now be simple strings which
defaults to `lambda item: "[{!r}]".format(exception) in item`.