fix: permission error when cleaning up fetches directory in run-task (#22)
This is an attempt to fix permission errors cropping up when cleaning up the fetches directory and cleaning the git checkout from run-task on windows. The code in the remove function comes, almost verbatim, from a Mozilla Central utility called mozfile. This utility is used to deal with this sort of issues on Mozilla Central.

See https://searchfox.org/mozilla-central/source/testing/mozbase/mozfile/mozfile/mozfile.py

Some of the behavior of the remove function is not covered in the unit tests. The reason is, to trigger the retry behavior, the test must be running under windows (otherwise we can't simulate the conditions that trigger this code path)

See https://searchfox.org/mozilla-central/source/testing/mozbase/mozfile/tests/test_move_remove.py

The reason the patch doesn't introduce the mozfile module as a dependency is that run-task should be kept portable and depend only on a Python interpreter.