fix: stop defaulting to a dummy 'TASKCLUSTER_ROOT_URL'
Currently if not specified, standalone taskgraph defaults to
"https://taskcluster.net". This isn't a real TC deployment, and worse
the server at taskcluser.net redirects unfound pages back to the main
page. So if this gets used we often download the HTML of this page which
causes bizarre and hard to diagnose errors.

Instead, this patch sets the default to `None` and a clear error message
is printed if it gets hit. Individual projects can set the default like
so:
https://searchfox.org/mozilla-central/source/taskcluster/gecko_taskgraph/__init__.py#19