ci: replace black, flake8 and yamllint tasks with pre-commit
We'll rely on https://pre-commit.ci to run the tasks. While using
Taskcluster for this is better ideologically, pre-commit.ci has some
compelling advantages:

1. The pre-commit cache is shared across all CI users so tasks will
almost never need to clone repos. Due to the low volume of commits to
taskgraph, if we used Taskcluster we'd rarely get a cache hit and tasks
would be *much* slower.
2. The pre-commit Github integration can automatically fix PRs.
3. The pre-commit Github integration can automatically submit PRs to
update dependencies.