fix: use 'zipfile' module to extract fetches on Windows
Zipfiles generated natively on Windows use backslash path separators for
their internal path structure. The `unzip` utility in Linux is unable to
interpret this and creates files with `\` in their name rather than
directories.

Even if a task is running on Windows, the `unzip` utility is bundled
with `mozilla-build` and runs in the shell there. So it is still unable
to extract the archive properly.

This patch switches to extracting with the `zipfile` stdlib rather than
the `unzip` utility when running from Windows.

Note that the bug can still exist if a task running on Linux attempts to
consume a zipfile created from a task on Windows.