feat!: make 'IndexSearch' strategy take status and expiration into account
BREAKING CHANGE: Replacement optimization strategies now take a
'deadline' argument.

When replacing a task we don't currently take into account the status or
expiry of the task. This means if a task failed it is still a valid
candidate for replacement.

More subtly, if a task expires before the deadline of one of the
dependents that need it, it might not be available by the time it is
needed.

In both cases, we should not treat the task as a suitable candidate for
replacement.

This syncs a change from Gecko.

Jira: RELENG-889