feat!: allow optimizing tasks that weren't explicitly requested
In certain cases it may be possible to optimize a task if all its
dependents have been optimized. However we don't want to do this if it
was "explicitly requested" (e.g, via the 'always_target' attribute or if
the user specified it).

This introduces a 'requested_tasks' argument to the optimization process
which will help it distinguish if a task is requested or not and whether
it is safe to optimize.

This syncs a change from Gecko.

Jira: RELENG-889