Bug 1784232 - part 7: Create `base_ref` parameter and let it default to default branch
The upcoming `base_rev` parameter is not always a trustworthy bit of data. I'm not saying Github[1] is lying but in 3 common circumstances, `base_rev` is not usable as is.

--- New head branch, but same code ---

The first one is: when you create a new branch - say `release/X.Y.Z`, Github sets `base_rev` to `0000000000000000000000000000000000000000`. We get this value even though the branch is 100% identical to the base branch (e.g.: `main`). Therefore, git-based systems (like Gitlab[2]) consider that every single file is new on this branch. This wastes CI resources since we're testing the same code but just on a different branch.

--- Same head branch name, but force-pushed commits ---

The second case is even more common in PRs: git-based systems make `base_rev` the former `HEAD` of a force-pushed branch. This means `base_rev` doesn't exist in the DAG. So we end up with errors like this one[3]:

```
INFO - Processing push for commit range ac97a75506f57cc3c6a2e5afc12ed14124a72634 -> ac7cfed5f0ec5ca58bdab9bb2196e8d8e653f41a
FATAL: Invalid revision range ac97a75506f57cc3c6a2e5afc12ed14124a72634..ac7cfed5f0ec5ca58bdab9bb2196e8d8e653f41a
```

--- Same head branch, but moving `base_rev` ---

Another common case during PRs: the base branch (usually `main`) will keep moving while a PR is open. This means, the `base_rev` won't be the actual base of a PR. So the revision range will be as busted as the one up above.

--- 3 problems, 1 solution ---

All in all, we have to determine a more accurate `base_rev` than the one Github provides because it's either too far behind causing a full rebuild, or it's non-existent, or it's not part of the ancestors of the `head_rev`. `base_rev` needs to be latest common revision between the base branch and the head one. In order to determine it, we need the `base_ref`. Otherwise in the case we get `000000000...`, the latest common revision turns out to be `00000000...`.

Hence, if the `base_ref` happens not to be provided, defaulting it to the default branch (usually `main`) is a good enough assumption, I believe. In fact, PRs are usually open against `main`/`master`.

[1] Gitlab actually behaves the same way which makes me think it's a git thing.
[2] https://gitlab.com/gitlab-org/gitlab/-/issues/15170
[3] https://firefox-ci-tc.services.mozilla.com/tasks/dPL42ewfTEekFXMeVaWuyA/runs/0/logs/public/logs/live.log#L214