ci: switch to the 'public_restricted' pull request policy
This means pull requests opened by non-collaborators will assume a
different role and have a different set of scopes associated to them.

For taskgraph, the only task we'd want to limit is the `codecov-upload`
one to avoid leaking the codecov API token. So we add a
`run-on-tasks-for` key here which will filter out tasks_for ==
"github-pull-request-untrusted".