refactor!: drop support for `always_target` attribute
BREAKING CHANGE

This is a feature that was cargo culted from Gecko, but likely isn't
useful or needed outside of Gecko. We are moving it out of
`generator.py` over there, and dropping the feature here.

If consumers do end up needing something like this, they can
re-implement it similar to how Gecko does. We also look into
re-introducing the functionality in core Taskgraph again if needed
(with nicer implementation).

Issue: #29