Bug 1621845 - Normalize fetch path in fetch-content.
The win64-aarch64 have a kind of a nasty trick that makes fetch-content
download artifacts of a dependent task directly as artifacts of the task
itself. For some reason, while this pattern works on native Windows
jobs, it doesn't on Linux. What happens is essentially that:

  `pathlib.Path(path).joinpath('../foo').mkdir(parents=True, exist=ok=True)`

fails when path doesn't exist first. I guess the fetches directory
already exists on Windows worker or something.

Unfortunately, os.path.normpath doesn't take `pathlib.Path`s in
still-supported python 3.5, so we have to convert to str first.