feat: Add ability to enable/disable the 'always_target' attribute via parameter
This syncs a change from Gecko.

It doesn't always make sense to use the `always_target` feature. E.g, in
Gecko, we only want to enable it for try and nowhere else. This allows
projects to disable it in certain instances. It defaults to `True` for
now to preserve backwards compatibility.