fix: make zstandard an optional dependency
It turns out we don't want to vendor zstandard into mozilla-central
because it requires compiled binaries.

A) The vendor results in 3.5Mb of unused binaries being added to the
Gecko source, and
B) zstandard is listed as a "pypi" dependency in the mach environment.
Also adding it as a "vendored" dependency causes some issues that need
to be worked around.

Overall it is simpler if Taskgraph only lists it as an optional
dependency so we don't end up pulling it into m-c.