fix: busted Python docker-image task
This was failing because pyenv introduced a native `latest` subcommand
which was interfering with the one from the xx-env plugin. The good news
is that now that this `latest` subcommand exists, xx-env is no longer
necessary.

This also adds `set -e` to the pyenv script so we don't miss failures in
the future.