Memoize taskgraph.util.load_yaml
It _feels_ like this ought to be safe to do everywhere, but it's difficult to shake the feeling that there may be some edge cases lurking out there where a file would _want_ to be reloaded. The only plausible way this seems like it _should_ happen is in a scenario where a yaml file is being written and loaded during taskgraph generation...which is pretty far fetched, and possibly even something we would explicitly not support. Any such scenario would reduce the determinism of taskgraph, which seems like something that should be avoided.

I did test in Gecko with a number of different parameter sets, and the taskgraphs came out identical with and without this memoized - so that's _some_ level of assurance that this is not totally out there.