Add smarter defaults for loaders, transforms, and treeherders (#217)
* Add a new default loader that uses the `job` and `task` transforms by default

This is introducing a new loader called `default`. This loader outsources all real loading work to the `transform` loader, and really only exists to be able to specify our new default transforms without breaking compabitility.

Default transforms could be a bit more debatable - but it seems to me that the vast majority of docker and generic worker tasks use both job and task transforms, so they seem like pretty sensible defaults.

This also fixes a small bug I found in `FakeKind`, where certain tests don't work in isolation because `loaded_tasks` may be undefined. (Many other tests actually modifying this class and set it, so running in a batch almost always works.)

Co-authored by jcristau.

* Add support for `treeherder: true` in tasks to get autogenerated treeherder metadata.

This is largely the same as what's described in #184 with a couple of small exceptions:
- There's no support for dealing with conflict symbols if two kinds share the same first letter -- implementing this turned out to be more work than I think it was worth, as we have to keep track of them globally. This would probably also preclude parallelizing loading/transforming.
- Platform defaults to `default/opt`. I would've liked to try to do something a bit better for the first part of this, but the only ideas I could come up with were `default` and using the full task label. `default` seemed better, because it avoid growing Treeherder vertically. (Once you get to the point of growing enough horizontally, you're probably ready to start defining your own metadata...)

There's a large `add_index_routes` transform getting moved around in this patch. No code has changed there, it just depends on `tier` being set, so we must have it happen after treeherder defaults are handled.