Rework treeherder defaults to fix indexes that get dropped sometimes.
It turns out that `build_task` needs to run after `add_index_routes` (big whoops on my part...).

Fixing this means extracting all the treeherder goop from `build_task` so we can run it before `add_index_routes`. In the process I noticed that my original patch ended up adding unnecessary `kind` and `platform` keys to the `extra.treeherder` object, which I'm also correcting here. (This was due to using the `treeherder` object as both an intermediary object to merge the defaults and task provided treeherder information and the final thing we set on the task - something we weren't doing before.)